+++ mayan/apps/checkouts/tests/test_links.py
@@ -1,6 +1,8 @@
 from __future__ import unicode_literals
 
+from mayan.apps.documents.permissions import permission_document_new_version
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
+from mayan.apps.sources.links import link_document_version_upload
 
 from ..links import link_check_out_document, link_check_out_info
 from ..permissions import (
@@ -47,3 +49,27 @@ class CheckoutLinksTestCase(
         )
         resolved_link = self._resolve_document_check_out_info_link()
         self.assertNotEqual(resolved_link, None)
+
+
+class DocumentVersionListViewTestCase(
+    DocumentCheckoutTestMixin, GenericDocumentViewTestCase
+):
+    def _get_document_new_version_link(self):
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_document_new_version
+        )
+
+        self.add_test_view(test_object=self.test_document)
+        context = self.get_test_view()
+        return link_document_version_upload.resolve(context=context)
+
+    def test_document_version_new_not_blocked(self):
+        resolved_link = self._get_document_new_version_link()
+        self.assertNotEqual(resolved_link, None)
+
+    def test_document_version_new_blocked(self):
+        self._check_out_test_document()
+
+        resolved_link = self._get_document_new_version_link()
+        self.assertEqual(resolved_link, None)