+++ mayan/apps/checkouts/tests/test_views.py
@@ -1,8 +1,11 @@
 from __future__ import unicode_literals
 
-from mayan.apps.documents.permissions import permission_document_view
+from mayan.apps.documents.models import DocumentVersion
+from mayan.apps.documents.permissions import (
+    permission_document_new_version, permission_document_version_view,
+    permission_document_view
+)
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
-from mayan.apps.sources.links import link_document_version_upload
 
 from ..literals import STATE_CHECKED_OUT, STATE_LABELS
 from ..models import DocumentCheckout
@@ -378,40 +381,26 @@ class NewVersionBlockViewTestCase(
     GenericDocumentViewTestCase
 ):
     def test_document_check_out_block_new_version(self):
-        """
-        Gitlab issue #231
-        User shown option to upload new version of a document even though it
-        is blocked by checkout - v2.0.0b2
-
-        Expected results:
-            - Link to upload version view should not resolve
-            - Upload version view should reject request
-        """
-        self._create_test_case_superuser()
         self._check_out_test_document()
-        self.login_superuser()
+        version_count = DocumentVersion.objects.count()
+
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_document_new_version
+        )
+        self.grant_access(
+            obj=self.test_document,
+            permission=permission_document_version_view
+        )
 
         response = self.post(
             viewname='sources:document_version_upload', kwargs={
-                'document_pk': self.test_document.pk
+                'document_id': self.test_document.pk
             }, follow=True
         )
-
         self.assertContains(
-            response=response, text='blocked from uploading',
+            response=response, text='Unable to upload new versions',
             status_code=200
         )
 
-        response = self.get(
-            viewname='documents:document_version_list', kwargs={
-                'pk': self.test_document.pk
-            }, follow=True
-        )
-
-        # Needed by the url view resolver
-        response.context.current_app = None
-        resolved_link = link_document_version_upload.resolve(
-            context=response.context
-        )
-
-        self.assertEqual(resolved_link, None)
+        self.assertEqual(DocumentVersion.objects.count(), version_count)